- added DAO and DTO files to business and order models - added lombok annotations for convinience - added some spring validation annotations
