--- src/main/java/de/berlin/lostberlin/model/Business.java
@@ -1,164 +1,0 @@
-package de.berlin.lostberlin.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import java.io.Serializable;
-import java.util.Date;
-
-@Entity
-@Table(name = "business")
-@EntityListeners(AuditingEntityListener.class)
-@JsonIgnoreProperties(value = {"createdAt", "updatedAt"}, allowGetters = true)
-
-public class Business implements Serializable {
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Long id;
-        @NotBlank
-        private String fName;
-        @NotBlank
-        private String lName;
-        @NotBlank
-        private String eMail;
-
-        private String phone;
-
-        private String description;
-
-        private String serviceLocation;
-
-        private byte[] photo;
-
-    private String login;
-
-    private String password;
-
-
-        @Column(nullable = false, updatable = false)
-        @Temporal(TemporalType.TIMESTAMP)
-        @CreatedDate
-        private Date createdAt;
-
-        @Column(nullable = false)
-        @Temporal(TemporalType.TIMESTAMP)
-        @LastModifiedDate
-        private Date updatedAt;
-
-    public Business()  {
-    }
-
-    public Business(Long id, @NotBlank String fName, @NotBlank String lName, @NotBlank String eMail, String phone, String description, String serviceLocation, byte[] photo, String login, String password, Date createdAt, Date updatedAt) {
-        this.id = id;
-        this.fName = fName;
-        this.lName = lName;
-        this.eMail = eMail;
-        this.phone = phone;
-        this.description = description;
-        this.serviceLocation = serviceLocation;
-        this.photo = photo;
-        this.login = login;
-        this.password = password;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getfName() {
-        return fName;
-    }
-
-    public void setfName(String fName) {
-        this.fName = fName;
-    }
-
-    public String getlName() {
-        return lName;
-    }
-
-    public void setlName(String lName) {
-        this.lName = lName;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return eMail;
-    }
-
-    public void setEmail(String eMail) {
-        this.eMail = eMail;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getServiceLocation() {
-        return serviceLocation;
-    }
-
-    public void setServiceLocation(String serviceLocation) {
-        this.serviceLocation = serviceLocation;
-    }
-
-    public byte[] getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(byte[] photo) {
-        this.photo = photo;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Date getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
