+++ src/main/java/de/berlin/lostberlin/model/Order/StatusTypes.java
@@ -1,0 +1,7 @@
+package de.berlin.lostberlin.model.Order;
+
+public enum StatusTypes {
+    PENDING,
+    CLOSED,
+    CONFIRMED
+}
